phase1: correctly exclude kmods files from rsync target upload and prune
There is currently a BUG in the --exclude pattern for the targetupload step.
The /kmods/ directory is actually never excluded. This was never notice
as the rsynclist never actually had kmods entry as make checksum step
was done before moving the kmods to the dedicated directory.

This is caused by the fact that with --files-from, not only the /kmods/
directory needs to be excluded but also every content in it, hence the
additional --exclude "/kmods/**" is required to correcly skip the entry
present in rsynclist.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>