phase1: fix CONFIG_SIGNED_PACKAGES for apk signing
Currently `CONFIG_SIGNED_PACKAGES` is only enabled when `usign_key` is
configured. However, we also want to enable it when `apk_key` is
configured, as we support signing APK packages as well.

So lets fix it by adding `IsSignedPackagesEnabled` helper which checks
for both keys and updates the .config generation to use it.

Signed-off-by: Petr Štetiar <ynezz@true.cz>